+++ www/apps/ui/src/examples/label-base-plus.tsx
@@ -1,0 +1,9 @@
+import { Label } from "@medusajs/ui"
+
+export default function LabelBasePlus() {
+  return (
+    <Label size="base" weight="plus">
+      Label
+    </Label>
+  )
+}